backend/requirements.txt
# Production dependencies — flexible pins for Python 3.9+ compatibility

fastapi==0.110.1
uvicorn>=0.25.0
starlette==0.37.2
python-multipart>=0.0.9
python-dotenv>=1.0.0

# Database
motor>=3.5.0
pymongo>=4.9.0
dnspython>=2.4.0

# Auth
bcrypt>=4.0.0
PyJWT>=2.0.0
passlib>=1.7.0

# Validation & serialization
pydantic>=2.0.0
email-validator>=2.0.0

# AWS S3
boto3>=1.34.0
botocore>=1.34.0

# PDF generation
reportlab>=4.0.0
pillow>=9.0.0
qrcode[pil]>=7.4.0

# AI — Google Gemini
google-generativeai>=0.8.0

# HTTP / async
anyio>=3.6.0
httpx>=0.27.0
certifi>=2023.0.0
requests>=2.28.0
idna>=3.0
urllib3>=1.26.0
pycryptodome>=3.20.0

# Utilities
typing_extensions>=4.0.0
click>=8.0.0
